+++ packages/dredd-transactions/scripts/prepack.js
@@ -12,7 +12,6 @@ const PACKAGE_DIR = path.resolve(__dirname, '..');
 const SYMLINKS_LOG = path.join(PACKAGE_DIR, 'prepack-symlinks.log');
 const APIB_PARSER_PATH = path.join(PACKAGE_DIR, 'node_modules', '@apielements', 'apib-parser');
 
-
 function readPackageJson(packageDir) {
   return JSON.parse(fs.readFileSync(path.join(packageDir, 'package.json')));
 }
@@ -65,7 +64,6 @@ function symlinkDependencyTreeToLocalNodeModules(dependencyName) {
   dependencies.forEach(symlinkDependencyTreeToLocalNodeModules);
 }
 
-
 // make sure all bundled deps are accessible in the local 'node_modules' dir
 const packageData = readPackageJson(PACKAGE_DIR);
 const { bundledDependencies } = packageData;
@@ -82,5 +80,5 @@ writePackageJson(APIB_PARSER_PATH, apibParserPackageData);
   path.join(PACKAGE_DIR, 'node_modules', 'protagonist'),
   path.join(PACKAGE_DIR, '..', '..', 'node_modules', 'protagonist'),
 ]
-  .filter(protagonistPath => fs.existsSync(protagonistPath))
-  .forEach(protagonistPath => rimraf.sync(protagonistPath));
+  .filter((protagonistPath) => fs.existsSync(protagonistPath))
+  .forEach((protagonistPath) => rimraf.sync(protagonistPath));