+++ packages/dredd/test/fixtures/sanitation/request-headers.js
@@ -2,7 +2,7 @@ const caseless = require('caseless');
 const hooks = require('hooks');
 
 hooks.after('Resource > Update Resource', (transaction, done) => {
-  const headers = transaction.test.request.headers;
+  const {headers} = transaction.test.request;
   const name = caseless(headers).has('Authorization');
   delete headers[name];
   transaction.test.request.headers = headers;